Refactoring and commenting.
